pyproject.toml
[project]
name = "faststream-request-reply"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.1",
    "faststream[cli,nats]>=0.6.7",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
